+++ android/app/src/main/java/com/app/edonymyeon/presentation/ui/alarmsetting/AlarmSettingViewModel.kt
@@ -1,0 +1,5 @@
+package com.app.edonymyeon.presentation.ui.alarmsetting
+
+import androidx.lifecycle.ViewModel
+
+class AlarmSettingViewModel : ViewModel()
