+++ android/app/src/main/res/values/colors.xml
@@ -14,4 +14,5 @@
     <color name="gray_434343">#434343</color>
     <color name="gray_615f5f">#615F5F</color>
     <color name="gray_d9d9d9">#d9d9d9</color>
+    <color name="gray_666666">#666666</color>
 </resources>
