+++ android/app/src/main/res/values/strings.xml
@@ -135,4 +135,10 @@
     <string name="sign_up_verify_duplication">중복확인</string>
     <string name="sign_up_nickname_policy">* 1~20자 이내로 가능합니다</string>
     <string name="sign_up_password_policy">* 8~30자, 알파벳, 숫자, 특수기호\n 모두 1개 이상 포함되어야 합니다.</string>
+
+    <!--alarm_setting-->
+    <string name="alarm_setting_tool_bar_title">푸시 알림 설정</string>
+    <string name="alarm_setting_title">푸시 알림 수신</string>
+    <string name="alarm_setting_content">해제하면 푸시 알림을 수신할 수 없습니다.</string>
+
 </resources>
